requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
